Remove dependency on TypeToken
